(ref_soft) 64bit patches to prevent crash at startup -
https://github.com/skullernet/q2pro/commit/330cbe5ea9921e2d653d7c6782a5a29e1f4ac263#diff-591b3269a4e1e2d0ffeb23471c8a6d8d

Also adding this for good measure -

https://github.com/skullernet/q2pro/commit/75540aa539866684f1056a2b6a1285a05a053f35#diff-a943ecedd8e0ee0e9b47dc27f6f14872

+++ ref_soft/r_main.c
@@ -323,8 +323,8 @@ qboolean SWR_Init( void *hInstance, void *wndProc )
 
 // TODO: collect 386-specific code in one place
 #if	id386
-	Sys_MakeCodeWriteable ((long)R_EdgeCodeStart,
-					     (long)R_EdgeCodeEnd - (long)R_EdgeCodeStart);
+	Sys_MakeCodeWriteable ((uintptr_t)R_EdgeCodeStart,
+					     (uintptr_t)R_EdgeCodeEnd - (uintptr_t)R_EdgeCodeStart);
 	Sys_SetFPCW ();		// get bit masks for FPCW	(FIXME: is this id386?)
 #endif	// id386
 
@@ -871,13 +871,13 @@ void R_EdgeDrawing (void)
 	else
 	{
 		r_edges =  (edge_t *)
-				(((long)&ledges[0] + CACHE_SIZE - 1) & ~(CACHE_SIZE - 1));
+				(((uintptr_t)&ledges[0] + CACHE_SIZE - 1) & ~(CACHE_SIZE - 1));
 	}
 
 	if (r_surfsonstack)
 	{
 		surfaces =  (surf_t *)
-				(((long)&lsurfs[0] + CACHE_SIZE - 1) & ~(CACHE_SIZE - 1));
+				(((uintptr_t)&lsurfs[0] + CACHE_SIZE - 1) & ~(CACHE_SIZE - 1));
 		surf_max = &surfaces[r_cnumsurfs];
 	// surface 0 doesn't really exist; it's just a dummy because index 0
 	// is used to indicate no edge attached to surface
